# ipx/docs/swagger.yaml
basePath: /
definitions:
  misc.BalanceRequest:
    properties:
      address:
        example: 0xEbD69375...
        type: string
      block:
        example: latest
        type: string
    type: object
  misc.BalanceResponse:
    properties:
      balance:
        type: string
      balance_hex:
        type: string
      ether:
        type: string
      gwei:
        type: string
      wei:
        type: string
    type: object
  misc.BaseFeePerGasResponse:
    properties:
      base_fee_per_gas:
        type: string
      base_fee_per_gas_hex:
        type: string
      ether:
        type: string
      gwei:
        type: string
      wei:
        type: string
    type: object
  misc.BlockNumberResponse:
    properties:
      block_number:
        type: integer
      block_number_hex:
        type: string
    type: object
  misc.CallRequest:
    properties:
      block:
        example: latest
        type: string
      data:
        example: 0x70a08231...
        type: string
      from:
        example: 0xEbD69375...
        type: string
      to:
        example: 0x8336c196...
        type: string
    type: object
  misc.ChainIDResponse:
    properties:
      chain_id:
        type: integer
      chain_id_hex:
        type: string
    type: object
  misc.CodeRequest:
    properties:
      address:
        example: 0xEbD69375...
        type: string
      block:
        example: latest
        type: string
    type: object
  misc.CodeResponse:
    properties:
      code:
        type: string
      is_contract:
        type: boolean
    type: object
...
  misc.EstimateGasRequest:
    properties:
      block:
        example: latest
        type: string
      data:
        example: 0x
        type: string
      from:
        example: 0xEbD69375...
        type: string
      to:
        example: 0x8336c196...
        type: string
      value:
        example: "1000000000000000000"
        type: string
    type: object
  misc.EstimateGasResponse:
    properties:
      ether:
        type: string
      gas_limit:
        type: integer
      gas_limit_hex:
        type: string
      gwei:
        type: string
      wei:
        type: string
    type: object
  misc.GasPriceResponse:
    properties:
      ether:
        type: string
      gas_price:
        type: string
      gas_price_hex:
        type: string
      gwei:
        type: string
      wei:
        type: string
    type: object
...
  misc.MaxFeePerGasResponse:
    properties:
      ether:
        type: string
      gwei:
        type: string
      max_fee_per_gas:
        type: string
      max_fee_per_gas_hex:
        type: string
      wei:
        type: string
    type: object
  misc.MaxPriorityFeePerGasResponse:
    properties:
      ether:
        type: string
      gwei:
        type: string
      max_priority_fee_per_gas:
        type: string
      max_priority_fee_per_gas_hex:
        type: string
      wei:
        type: string
    type: object
  misc.NonceRequest:
    properties:
      address:
        example: 0xEbD69375...
        type: string
      block:
        example: pending
        type: string
    type: object
  misc.NonceResponse:
    properties:
      nonce:
        type: integer
      nonce_hex:
        type: string
    type: object
  misc.RawRPCRequest:
    properties:
      method:
        type: string
      params:
        items: {}
        type: array
    type: object
  misc.RawRPCResponse:
    properties:
      result: {}
    type: object
...
  misc.SendTransactionRequest:
    properties:
      raw_tx:
        example: 0x02f8...
        type: string
    type: object
  misc.SendTransactionResponse:
    properties:
      tx_hash:
        type: string
    type: object
...
  misc.TransactionRequest:
    properties:
      tx_hash:
        example: 0xabc123...
        type: string
    type: object
...
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
...
  /evm/rpc:
    post:
      consumes:
      - application/json
      description: Proxies a raw JSON-RPC call to the connected node and returns the
        result as-is
      parameters:
      - description: JSON-RPC method and params
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.RawRPCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.RawRPCResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Call any JSON-RPC method
      tags:
      - rpc
  /evm/rpc/balance:
    post:
      consumes:
      - application/json
      description: Returns the balance for the given address at the specified block
        in decimal and hex
      parameters:
      - description: Address and block
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.BalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.BalanceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get balance for address
      tags:
      - rpc
  /evm/rpc/block-number:
    post:
      description: Returns the latest block number in decimal and hex
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.BlockNumberResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get block number
      tags:
      - rpc
  /evm/rpc/call:
    post:
      consumes:
      - application/json
      description: Executes eth_call and returns the raw result
      parameters:
      - description: Call parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.CallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Execute a read-only contract call
      tags:
      - rpc
  /evm/rpc/chain-id:
    post:
      description: Returns the chain ID of the connected network as a decimal integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.ChainIDResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get chain ID
      tags:
      - rpc
  /evm/rpc/code:
    post:
      consumes:
      - application/json
      description: Returns the deployed code for the given address and whether it
        is a contract
      parameters:
      - description: Address and block
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.CodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.CodeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get code for address
      tags:
      - rpc
  /evm/rpc/fee/base:
    post:
      description: Returns the latest block baseFeePerGas in decimal and hex
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.BaseFeePerGasResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get base fee per gas
      tags:
      - rpc
  /evm/rpc/fee/max:
    post:
      description: Returns max_fee_per_gas calculated as base_fee_per_gas * 2 + max_priority_fee_per_gas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.MaxFeePerGasResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get max fee per gas
      tags:
      - rpc
  /evm/rpc/fee/priority:
    post:
      description: Returns the recommended priority fee per gas in decimal and hex
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.MaxPriorityFeePerGasResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get max priority fee per gas
      tags:
      - rpc
  /evm/rpc/gas/estimate:
    post:
      consumes:
      - application/json
      description: Returns the estimated gas limit for the given transaction parameters
      parameters:
      - description: Transaction parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.EstimateGasRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.EstimateGasResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Estimate gas for a transaction
      tags:
      - rpc
  /evm/rpc/gas/price:
    post:
      description: Returns the current gas price in decimal and hex
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.GasPriceResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get gas price
      tags:
      - rpc
  /evm/rpc/nonce:
    post:
      consumes:
      - application/json
      description: Returns the pending transaction count (nonce) for the given address
        as a decimal integer
      parameters:
      - description: Address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.NonceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.NonceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get nonce for address
      tags:
      - rpc
  /evm/rpc/transaction:
    post:
      consumes:
      - application/json
      description: Returns transaction details for the given tx hash
      parameters:
      - description: Transaction hash
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get transaction by hash
      tags:
      - rpc
  /evm/rpc/transaction/receipt:
    post:
      consumes:
      - application/json
      description: Returns the transaction receipt for the given tx hash
      parameters:
      - description: Transaction hash
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get transaction receipt by hash
      tags:
      - rpc
  /evm/rpc/transaction/send:
    post:
      consumes:
      - application/json
      description: Broadcasts a signed raw transaction with eth_sendRawTransaction
      parameters:
      - description: Signed raw transaction
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.SendTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.SendTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send signed raw transaction
      tags:
      - rpc
...
swagger: "2.0"